+++ _examples/configuration/from-yaml-file/main.go
@@ -12,6 +12,8 @@ func main() {
 	// [...]
 
 	// Good when you have two configurations, one for development and a different one for production use.
+	// If iris.YAML's input string argument is "~" then it loads the configuration from the home directory
+	// and can be shared between many iris instances.
 	app.Run(iris.Addr(":8080"), iris.WithConfiguration(iris.YAML("./configs/iris.yml")))
 
 	// or before run: