+++ core/maintenance/maintenance.go
@@ -1,0 +1,6 @@
+package maintenance
+
+// Start starts the maintenance process.
+func Start(globalConfigurationExisted bool) {
+	CheckForUpdates(!globalConfigurationExisted)
+}