+++ core/router/router_spa_wrapper.go
@@ -1,7 +1,6 @@
 package router
 
 import (
-	"net/http"
 	"strings"
 
 	"github.com/kataras/iris/context"
@@ -19,8 +18,17 @@ type SPABuilder struct {
 	AssetValidators []AssetValidator
 }
 
+// AddIndexName will add an index name.
+// If path == $filename then it redirects to "/".
+//
+// It can be called after the `BuildWrapper ` as well but BEFORE the server start.
+func (s *SPABuilder) AddIndexName(filename string) *SPABuilder {
+	s.IndexNames = append(s.IndexNames, filename)
+	return s
+}
+
 // NewSPABuilder returns a new Single Page Application builder
-// It does what StaticWeb expected to do when serving files and routes at the same time
+// It does what StaticWeb or StaticEmbedded expected to do when serving files and routes at the same time
 // from the root "/" path.
 //
 // Accepts a static asset handler, which can be an app.StaticHandler, app.StaticEmbeddedHandler...
@@ -32,11 +40,13 @@ func NewSPABuilder(assetHandler context.Handler) *SPABuilder {
 	}
 
 	return &SPABuilder{
-		IndexNames:   []string{"index.html"},
+		IndexNames: nil,
+		// IndexNames is empty by-default,
+		// if the user wants to redirect to "/" from "/index.html" she/he can chage that to  []string{"index.html"} manually.
 		AssetHandler: assetHandler,
 		AssetValidators: []AssetValidator{
 			func(path string) bool {
-				return strings.Contains(path, ".")
+				return true // returns true by-default, if false then it fires 404.
 			},
 		},
 	}
@@ -51,46 +61,31 @@ func (s *SPABuilder) isAsset(reqPath string) bool {
 	return true
 }
 
-// BuildWrapper returns a wrapper which serves the single page application
-// with the declared configuration.
-//
-// It should be passed to the router's `WrapRouter`:
-// https://godoc.org/github.com/kataras/iris/core/router#Router.WrapRouter
-//
-// Example: https://github.com/kataras/iris/tree/master/_examples/file-server/single-page-application-builder
-func (s *SPABuilder) BuildWrapper(cPool *context.Pool) WrapperFunc {
+// Handler serves the asset handler but in addition, it makes some checks before that,
+// based on the `AssetValidators` and `IndexNames`.
+func (s *SPABuilder) Handler(ctx context.Context) {
+	path := ctx.Path()
 
-	fileServer := s.AssetHandler
-	indexNames := s.IndexNames
-
-	wrapper := func(w http.ResponseWriter, r *http.Request, router http.HandlerFunc) {
-		path := r.URL.Path
-
-		if !s.isAsset(path) {
-			// it's not asset, execute the registered route's handlers
-			router(w, r)
-			return
-		}
-
-		ctx := cPool.Acquire(w, r)
-
-		for _, index := range indexNames {
-			if strings.HasSuffix(path, index) {
-				localRedirect(ctx, "./")
-				cPool.Release(ctx)
-				// "/" should be manually  registered.
-				// We don't setup an index handler here,
-				// let full control to the user
-				// (use middleware, ctx.ServeFile or ctx.View and so on...)
-				return
-			}
-		}
-
-		// execute file server for path
-		fileServer(ctx)
-
-		cPool.Release(ctx)
+	// make a validator call, by-default all paths are valid and this codeblock doesn't mean anything
+	// but for cases that users wants to bypass an asset she/he can do that by modifiying the `APIBuilder#AssetValidators` field.
+	//
+	// It's here for backwards compatibility as well, see #803.
+	if !s.isAsset(path) {
+		// it's not asset, execute the registered route's handlers
+		ctx.NotFound()
+		return
 	}
 
-	return wrapper
+	for _, index := range s.IndexNames {
+		if strings.HasSuffix(path, index) {
+			localRedirect(ctx, "./")
+			// "/" should be manually  registered.
+			// We don't setup an index handler here,
+			// let full control to the user
+			// (use middleware, ctx.ServeFile or ctx.View and so on...)
+			return
+		}
+	}
+
+	s.AssetHandler(ctx)
 }