+++ sessions/sessiondb/redis/database.go
@@ -12,7 +12,6 @@ import (
 // Database the redis back-end session database for the sessions.
 type Database struct {
 	redis *service.Service
-	async bool
 }
 
 // New returns a new redis database.
@@ -27,10 +26,9 @@ func (db *Database) Config() *service.Config {
 	return db.redis.Config
 }
 
-// Async if true passed then it will use different
-// go routines to update the redis storage.
+// Async is DEPRECATED
+// if it was true then it could use different to update the back-end storage, now it does nothing.
 func (db *Database) Async(useGoRoutines bool) *Database {
-	db.async = useGoRoutines
 	return db
 }
 
@@ -70,11 +68,7 @@ func (db *Database) Load(sid string) (storeDB sessions.RemoteStore) {
 
 // Sync syncs the database.
 func (db *Database) Sync(p sessions.SyncPayload) {
-	if db.async {
-		go db.sync(p)
-	} else {
-		db.sync(p)
-	}
+	db.sync(p)
 }
 
 func (db *Database) sync(p sessions.SyncPayload) {